refactor(core): Move TokenCounter to metrics directory
- Move TokenCounter class and tests to metrics directory

- Update import paths in related files

- Improve organization of token counting related code
