+++ src/core/metrics/TokenCounter.ts
