+++ tests/core/metrics/TokenCounter.test.ts
@@ -1,6 +1,6 @@
 import { type Tiktoken, get_encoding } from 'tiktoken';
 import { type Mock, afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { TokenCounter } from '../../../src/core/tokenCount/tokenCount.js';
+import { TokenCounter } from '../../../src/core/metrics/TokenCounter.js';
 import { logger } from '../../../src/shared/logger.js';
 
 vi.mock('tiktoken', () => ({
