+++ tests/core/metrics/calculateMetrics.test.ts
@@ -2,7 +2,7 @@ import { type Mock, describe, expect, it, vi } from 'vitest';
 import type { ProcessedFile } from '../../../src/core/file/fileTypes.js';
 import { calculateAllFileMetrics } from '../../../src/core/metrics/calculateAllFileMetrics.js';
 import { calculateMetrics } from '../../../src/core/metrics/calculateMetrics.js';
-import { TokenCounter } from '../../../src/core/tokenCount/tokenCount.js';
+import { TokenCounter } from '../../../src/core/metrics/TokenCounter.js';
 import type { RepomixProgressCallback } from '../../../src/shared/types.js';
 import { createMockConfig } from '../../testing/testUtils.js';
